+++ NEWS.md
@@ -1,3 +1,8 @@
+# Release 0.0.0.9012
+
+- adjust factor levels for number of inserts (#4)
+- summarize number of inserts by transcript (#5)
+
 # Release 0.0.0.9011
 
 - fix order of factors